api/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
google-cloud-pubsub==2.18.0
pg8000==1.24.0
structlog==21.5.0
requests==2.31.0
google-api-python-client==2.94.0
ratelimit==2.2.1
backoff==2.1.2
dynaconf==3.1.9
pyjwt[crypto]==2.3.0